Fix: 댓글 응답 파일 분리, 스웨거 스키마 추가
